Remove FunctionClause error test
I am pretty sure this only existed to remind me that it's called
`:secret`. With the new type checking in the compiler this generates an
expected and spammy error... and it would remind me if I tried to use it
another way.

This being a relatively popular package, I bump the patch level to avoid
annoying someone somewhere with needless scroll.